requirements.txt
Django==3.1.3
django-cleanup==5.1.0
djangorestframework==3.12.2
Markdown==3.3.3
Pillow==8.0.1
